add methods #in_order, #pre_order, #post_order
will traverse binary search tree in respective order.

if block is given along with call of respective method, each
node will be handed off to the block with #yield. If no block is given
then an array of all nodes will be returned when traversal is complete.

more can be read in link below about the concept of inorder, preorder
and postorder traversal.

https://csanim.com/tutorials/inorder-preorder-and-postorder-tree-traversals-animated-guide